refactor: Change implementation of configuring node
This commit changes how a new node is configured. The
configuration is once again handled by interface_slurmd.
interface_slurmd now has a descriptor for accessing the
new_node status.

This commit also updates the BlockedStatus message shown
when slurmd is not running to "slurmd not running". This
is more descript than slurmd off which does not indicate
as well that the slurmd service is not running in the
background.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>