Remove etcd from slurmd operator (#21)
* chore(deps): Bump ops version to 2.*

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* chore(deps): Add systemd charm lib v1.3

Signed-off-by: NucciTheBoss <jason.nucciarone@canonical.com>

* refactor: Remove etcd from slurmd

Remove the dependency on etcd from slurmd. slurmd
now uses a systemd notices daemon to notify the
slurmd operator when the internal slurmd daemon
has successfully started.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>
Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* tests: Update unit tests

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* tests: Update integration tests

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* bugfix: Fix `node-configured` action

`node-configured` action was broken because the
method `ensure_slurmd_starts` was removed from
SlurmdCharm, but it was still being called by the
Slurmd integration interface. This action was fixed
by porting the interface_slurmd code over to the charm
using new utility modules added.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* refactor: Chnage module implementation

This commit implemenets feedback received for the
slurmd utility module. It changes the following:

* Switches write out from `override.conf` to 99-slurmd-charm.conf.
* Changes back-end implementation of restart() to use service_restart.
* Fixes comment to say "~30 sec" instead of "10 sec".

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* refactor: Change implementation of configuring node

This commit changes how a new node is configured. The
configuration is once again handled by interface_slurmd.
interface_slurmd now has a descriptor for accessing the
new_node status.

This commit also updates the BlockedStatus message shown
when slurmd is not running to "slurmd not running". This
is more descript than slurmd off which does not indicate
as well that the slurmd service is not running in the
background.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* tests: Update unit tests

Updates include the following:

* Update test_service_slurmd_stopped to evaluate new status message "slurmd not running".
* Update comment in test_update_status_success to be more descript.
* Update test_update_status_success to also evaluate the return value of _check_status.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* tests: Remove local charm builds from integration tests

The local charm builds were removed because it is preferred
that this code is contributed as a separate merge request.

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

* tests: Update unit test to not assume default status supplied by ops.testing.Harness

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>

---------

Signed-off-by: Jason C. Nucciarone <jason.nucciarone@canonical.com>
Signed-off-by: NucciTheBoss <jason.nucciarone@canonical.com>
Signed-off-by: Billy Olsen <billy.olsen@canonical.com>
Co-authored-by: Billy Olsen <billy.olsen@canonical.com>